% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{ABM_NI}
\alias{ABM_NI}
\alias{fweights}
\alias{distances}
\alias{selectEgo}
\alias{selectAlter}
\alias{opdelta}
\alias{opupdate}
\title{NI}
\usage{
ABM_NI(
  opinions,
  groups,
  net = NULL,
  H = 0.5,
  selectType = 1,
  prob = NULL,
  selectTypeAlter = 1,
  distance = NULL,
  iter,
  keep = TRUE,
  seed = NULL,
  verbose = TRUE
)

fweights(opinions, groups, H)

distances(opinions, groups, H)

selectEgo(nagents, selectType = 1, prob = NULL)

selectAlter(
  nagents,
  ego,
  net = NULL,
  distance = NULL,
  selectTypeAlter = 1,
  prob = NULL
)

opdelta(ego, alter, opinions, simweights)

opupdate(ego, opinions, delta)
}
\arguments{
\item{opinions}{Discrete variable; \code{min(opinion) = -1} and \code{max(opinion) = 1} with 20 steps.}

\item{groups}{Integer \{-1,1\}}

\item{net}{Matrix, the adjacency matrix representing the relations between
agents.  Valid values are 0 and 1.}

\item{H}{numeric $[0,1]$, relative influence of opinion (max if H = 0) and
group (max if H = 1)}

\item{selectType}{numeric \{1,2\}, Determines how egos are sampled. See
Details.}

\item{prob}{numeric vector, egos are sampled based on probabilities}

\item{selectTypeAlter}{numeric \{1,2,3,4,5,6\}, Determines how alters are
sampled. See Details.}

\item{distance}{numeric matrix. Higher scores indicate that dyads are
closer. See Details.}

\item{iter}{Integer. Number of pushes in each simulation run.}

\item{keep}{Logical. If \code{TRUE} complete chain is saved. If \code{FALSE} only
final result is saved.}

\item{seed}{Integer. Allows replication run.}

\item{verbose}{Logical. If \code{TRUE} returns results of each iteration on
screen.}

\item{nagents}{integer, number of agents}

\item{ego}{integer vector, selected ego(s) to receive push}

\item{alter}{integer vector, selected alter(s) to give push}

\item{simweights}{matrix, the similarity scores of the ego-alter pairs
calculated by \code{fweights}}

\item{delta}{numeric vector, the push(es) of alter(s)}
}
\value{
list of all kind of goodies.
}
\description{
Function to perform a NI ABM with several implementation
options....
}
\details{
For examples on how to use ... \code{selectType = 1}: only one ego;
\code{selectType = 2}: all egos put in a queue \code{probs}: sample probability for
ego(s) \code{selectTypeAlter = 1}, 1 random alter \code{selectTypeAlter = 2}, 1
random connected alter \code{selectTypeAlter = 3}, 1 alter, probability based on
similarity/distance \code{selectTypeAlter = 4}, 1 connected alter (path length
1), probability based on similarity/distance \code{selectTypeAlter = 5}, all
connected alters (path length 1) \code{selectTypeAlter = 6}, all alters
\code{distance}: If alters are selected based on distance between the ego-alter
pair this distance (or rather proximity) matrix is used. Higher scores
means higher probability to be selected. If distance plays are role, but
parameter is left to \code{NULL} distances based on similarity of opinion and
group is being used: $w_\{ij\} + 1$.
}
\examples{
seeds <- sample(2345:34567, 1000, replace = FALSE)
nagents <- 10
opinions <- c(-0.1, 1.0, 0.4, 0.4, -0.5, 0.4, 0.7, 0.7, -0.3, 0.6)
groups <- c(-1, 1, -1, 1, -1, 1, -1, 1, -1, 1)
net1 <- matrix(sample(c(0, 1), 100, replace = TRUE), nrow = 10, ncol = 10)
diag(net1) <- 0
}
\concept{NI}
